+++ src/DrawioAppServer/DrawioAppServer.ts
@@ -1,6 +1,0 @@
-import { Webview } from "vscode";
-import { DrawioInstance } from "../DrawioInstance";
-
-export interface DrawioAppServer {
-	setupWebview(webview: Webview): Promise<DrawioInstance>;
-}