+++ src/DrawioInstance.ts
@@ -23,7 +23,7 @@ export class DrawioInstance {
 
 	constructor(
 		private readonly messageStream: MessageStream,
-		private readonly options: { compressXml: boolean }
+		private readonly getConfig: () => Promise<DrawioConfig>
 	) {
 		this.dispose.track(
 			messageStream.registerMessageHandler((msg) =>
@@ -62,7 +62,7 @@ export class DrawioInstance {
 		});
 	}
 
-	private handleEvent(msg: DrawioEvent): void {
+	private async handleEvent(msg: DrawioEvent): Promise<void> {
 		if (msg.event === "init") {
 			this.onInitEmitter.emit();
 		} else if (msg.event === "autosave") {
@@ -88,11 +88,10 @@ export class DrawioInstance {
 			}
 			// do nothing
 		} else if (msg.event === "configure") {
+			const config = await this.getConfig();
 			this.sendAction({
 				action: "configure",
-				config: {
-					compressXml: this.options.compressXml,
-				},
+				config,
 			});
 		} else {
 			this.onUnknownMessageEmitter.emit({ message: msg });
@@ -296,26 +295,185 @@ type DrawioAction =
 			config: DrawioConfig;
 	  };
 
-interface DrawioConfig {
-	compressXml?: boolean;
+// See https://desk.draw.io/support/solutions/articles/16000058316-how-to-configure-draw-io-
+export interface DrawioConfig {
+	/**
+	 * An array of font family names in the format panel font drop-down list.
+	 */
+	defaultFonts?: string[];
+
+	/**
+	 * An array of font family names to be added before defaultFonts (9.2.4 and later).
+	 * Note: Fonts must be installed on the server and all client devices, or be added using the fontCss option. (6.5.4 and later).
+	 */
+	customFonts?: string[];
+
+	/**
+	 * Colour codes for the upper palette in the colour dialog (no leading # for the colour codes).
+	 */
+	presetColors?: string[];
+
+	/**
+	 * Colour codes to be added before presetColors (no leading # for the colour codes) (9.2.5 and later).
+	 */
+	customPresetColors?: string[];
+
+	/**
+	 * Available colour schemes in the style section at the top of the format panel (use leading # for the colour codes).
+	 * Possible colour keys are fill, stroke, gradient and font (font is ignored for connectors).
+	 */
+	defaultColorSchemes?: string[];
+
+	/**
+	 * Colour schemes to be added before defaultColorSchemes (9.2.4 and later).
+	 */
+	customColorSchemes?: string[];
+
+	/**
+	 * Defines the initial default styles for vertices and edges (connectors).
+	 * Note that the styles defined here are copied to the styles of new cells, for each cell.
+	 * This means that these values override everything else that is inherited from other styles or themes
+	 * (which may be supported at a later time).
+	 * Therefore, it is recommended to use a minimal set of values for the default styles.
+	 * To find the key/value pairs to be used, set the style in the application and find the key and value via Edit Style (Ctrl+E) (6.5.2 and later).
+	 * For example, to assign a default fontFamily of Courier New to all edges and vertices (and override all other default styles),
+	 * use
+	 * ```json
+	 * {
+	 * 	"defaultVertexStyle": {"fontFamily": "Courier New"},
+	 * 	"defaultEdgeStyle": {"fontFamily": "Courier New"}
+	 * }
+	 * ```
+	 * (6.5.2 and later).
+	 */
+	defaultVertexStyle?: Record<string, string>;
+
+	/**
+	 * See `defaultVertexStyle`.
+	 */
+	defaultEdgeStyle?: Record<string, string>;
+
+	/**
+	 * Defines a string with CSS rules to be used to configure the diagrams.net user interface.
+	 * For example, to change the background colour of the menu bar, use the following:
+	 * ```css
+	 * .geMenubarContainer { background-color: #c0c0c0 !important; }
+	 * .geMenubar { background-color: #c0c0c0 !important; }
+	 * ```
+	 * (6.5.2 and later).
+	 */
+	css?: string;
+
+	/**
+	 * Defines a string with CSS rules for web fonts to be used in diagrams.
+	 */
+	fontCss?: string;
+
+	/**
+	 * Defines a semicolon-separated list of library keys (unique names)
+	 * in a string to be initially displayed in the left panel (e.g. "general;uml;company-graphics").
+	 * Possible keys include custom entry IDs from the libraries field,
+	 * or keys for the libs URL parameter (6.5.2 and later).
+	 * The default value is `"general;uml;er;bpmn;flowchart;basic;arrows2"`.
+	 */
 	defaultLibraries?: string;
-	libraries?: {
+
+	/**
+	 * Defines an array of objects that list additional libraries and sections
+	 * in the left panel and the More Shapes dialog.
+	 */
+	libraries?: DrawioLibrarySection[];
+
+	/**
+	 * Defines the XML for blank diagrams and libraries (6.5.4 and later).
+	 */
+	emptyDiagramXml?: string;
+
+	/**
+	 * Specifies if the XML output should be compressed. The default is true.
+	 */
+	compressXml?: boolean;
+}
+
+interface DrawioLibrarySection {
+	title: DrawioResource;
+	entries: {
+		id: string;
+		preview?: string;
 		title: DrawioResource;
-		entries: {
-			id: string;
-			preview?: string;
+		desc?: DrawioResource;
+		libs: ({
 			title: DrawioResource;
-			desc?: DrawioResource;
-			libs: ({
-				title: DrawioResource;
-				tags?: string;
-			} & ({ data: string } | { url: string }))[];
-		}[];
+			tags?: string;
+		} & ({ data: unknown } | { url: string }))[];
 	}[];
 }
 
-interface DrawioResource {
+export interface DrawioLibraryData {
+	entryId: string;
+	libName: string;
+	data: { kind: "value"; value: unknown } | { kind: "url"; url: string };
+}
+
+export function simpleDrawioLibrary(
+	libs: DrawioLibraryData[]
+): DrawioLibrarySection[] {
+	function mapLib(lib: DrawioLibraryData) {
+		return lib.data.kind === "value"
+			? {
+					title: res(lib.libName),
+					data: lib.data.value,
+			  }
+			: {
+					title: res(lib.libName),
+					url: lib.data.url,
+			  };
+	}
+
+	const groupedLibs = groupBy(libs, (l) => l.entryId);
+
+	return [
+		{
+			title: res("Custom Libraries"),
+			entries: [...groupedLibs.values()].map((group) => ({
+				title: res(group.key),
+				id: group.key,
+				libs: group.items.map(mapLib),
+			})),
+		},
+	];
+}
+
+function res(name: string): DrawioResource {
+	return {
+		main: name,
+	};
+}
+
+export interface DrawioResource {
 	main: string;
 }
 
-type DrawioFormat = "html" | "xmlpng" | "png" | "xml" | "xmlsvg";
+export type DrawioFormat = "html" | "xmlpng" | "png" | "xml" | "xmlsvg";
+
+interface Group<TKey, TItem> {
+	key: TKey;
+	items: TItem[];
+}
+
+export function groupBy<TKey, T>(
+	items: ReadonlyArray<T>,
+	selectKey: (item: T) => TKey
+): Map<TKey, Group<TKey, T>> {
+	const map = new Map<TKey, Group<TKey, T>>();
+	for (const item of items) {
+		const key = selectKey(item);
+		let group = map.get(key);
+		if (!group) {
+			group = { key, items: [] };
+			map.set(key, group);
+		}
+		group.items.push(item);
+	}
+	return map;
+}